+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [2.1.0] - 2026-5-14
+
+### Removed
+- All obsolete methods from 2.0.0
+
+### Tooling
+- Improved documentation
+- Added missing tests
+
 ## [2.0.0] - 2026-4-14
 
 Major rework of the Failure system.